+++ CHANGELOG.rst
@@ -1,6 +1,19 @@
 Release notes
 =============
 
+Version v40.0.2
+---------------------
+- Update VulnerableCode documentation for advisories
+
+
+Version v40.0.1
+---------------------
+
+- Add throttling info in API docs, remove V1 and V2 API docs.
+- Avoid converting None to "None" in AdvisorySeverity.
+- Refine package curation workflow.
+
+
 Version v40.0.0
 ---------------------
 
@@ -36,7 +49,6 @@ Version v39.0.0
 - Add API/ UI support for Patch/PackageCommitPatch
 - Fix failing V2 pipelinea
 - Improve altcha challenge flow and reduce session timeout (https://github.com/aboutcode-org/vulnerablecode/pull/2348)
-
 
 Version v38.6.0
 ---------------------